Run fine-tune in background
