+++ fine_tune/app/qa_aws.py
@@ -2,15 +2,8 @@
 from fastapi import FastAPI, HTTPException
 from pydantic import BaseModel
 from inference.aws_inference import AwsInference
-from threading import Thread
-import asyncio
+from app.qa_app_factory import create_app
 
-app = FastAPI()
-
-# FastAPI schema and route
-class QARequest(BaseModel):
-    question: str
-    context: str
 
 S3_BUCKET = "opengpt2documents"
 # Set up AWS inference pipeline
@@ -20,27 +13,4 @@ inference = AwsInference(
     model_output_dir=f"s3://{S3_BUCKET}/checkpoints/finetuned-model"
 )
 
-
-@app.on_event("startup")
-async def startup_event():
-    # Run the blocking initialize() in a background thread without blocking the event loop
-    await asyncio.to_thread(inference.initialize)
-
-@app.post("/query")
-def answer_question(payload: QARequest):
-    if not inference.is_ready():
-        raise HTTPException(status_code=503, detail="Inference model not available.")
-
-    answer, score = inference.generate(payload.question, payload.context)
-    return {"answer": answer, "score": score}
-
-@app.get("/status")
-def get_status():
-    return {"status": inference.get_status()}
-
-@app.get("/")
-def root():
-    return {"message": "Fine-tune API is running."}
-
-
-
+app = create_app(inference)
