+++ fine_tune/app/qa_local.py
@@ -1,14 +1,9 @@
 # app/qa_local.py
 from fastapi import FastAPI, HTTPException
 from pydantic import BaseModel
+from app.qa_app_factory import create_app
 from inference.local_inference import LocalInference
 
-app = FastAPI()
-
-# FastAPI schema and route
-class QARequest(BaseModel):
-    question: str
-    context: str
 
 # Set up local inference pipeline
 inference = LocalInference(
@@ -17,18 +12,4 @@ inference = LocalInference(
     model_output_dir="./checkpoints"
 )
 
-inference.initialize()
-
-@app.post("/query")
-def answer_question(payload: QARequest):
-    if not inference.is_ready():
-        raise HTTPException(status_code=503, detail="Inference model not available.")
-
-    answer, score = inference.generate(payload.question, payload.context)
-    return {"answer": answer, "score": score}
-
-@app.get("/")
-def root():
-    return {"message": "Fine-tune API is running."}
-
-
+app = create_app(inference)
